requirements.txt
asgiref==3.4.1
Django==3.2.11
django-filter==21.1
djangorestframework==3.13.1
html5lib==1.1
importlib-metadata==4.10.1
Markdown==3.3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
zipp==3.7.0
psycopg2>=2.8

